extract Argument value logging, put impls in functional package
