+++ core/src/main/java/org/jdbi/v3/core/argument/EnumArgumentFactory.java
@@ -15,7 +15,7 @@ package org.jdbi.v3.core.argument;
 
 import java.lang.reflect.Type;
 import java.util.Optional;
-import org.jdbi.v3.core.argument.internal.LoggableToStringOrNPEArgument;
+import org.jdbi.v3.core.argument.internal.strategies.LoggableToStringOrNPEArgument;
 import org.jdbi.v3.core.config.ConfigRegistry;
 
 class EnumArgumentFactory implements ArgumentFactory {
