+++ core/src/main/java/org/jdbi/v3/core/argument/JavaTimeZoneIdArgumentFactory.java
@@ -13,7 +13,7 @@
  */
 package org.jdbi.v3.core.argument;
 
-import org.jdbi.v3.core.argument.internal.LoggableToStringOrNPEArgument;
+import org.jdbi.v3.core.argument.internal.strategies.LoggableToStringOrNPEArgument;
 import org.jdbi.v3.core.config.ConfigRegistry;
 
 import java.sql.Types;
