+++ postgres/src/main/java/org/jdbi/v3/postgres/DurationArgumentFactory.java
@@ -18,7 +18,7 @@ import java.time.Duration;
 
 import org.jdbi.v3.core.argument.AbstractArgumentFactory;
 import org.jdbi.v3.core.argument.Argument;
-import org.jdbi.v3.core.argument.internal.LoggableSetObjectArgument;
+import org.jdbi.v3.core.argument.internal.strategies.LoggableSetObjectArgument;
 import org.jdbi.v3.core.config.ConfigRegistry;
 import org.postgresql.util.PGInterval;
