+++ postgres/src/main/java/org/jdbi/v3/postgres/InetArgumentFactory.java
@@ -15,7 +15,7 @@ package org.jdbi.v3.postgres;
 
 import org.jdbi.v3.core.argument.AbstractArgumentFactory;
 import org.jdbi.v3.core.argument.Argument;
-import org.jdbi.v3.core.argument.internal.LoggableSetObjectArgument;
+import org.jdbi.v3.core.argument.internal.strategies.LoggableSetObjectArgument;
 import org.jdbi.v3.core.config.ConfigRegistry;
 
 import java.net.InetAddress;
