+++ postgres/src/main/java/org/jdbi/v3/postgres/UUIDArgumentFactory.java
@@ -18,7 +18,7 @@ import java.util.UUID;
 
 import org.jdbi.v3.core.argument.AbstractArgumentFactory;
 import org.jdbi.v3.core.argument.Argument;
-import org.jdbi.v3.core.argument.internal.LoggableSetObjectArgument;
+import org.jdbi.v3.core.argument.internal.strategies.LoggableSetObjectArgument;
 import org.jdbi.v3.core.config.ConfigRegistry;
 
 /**
